codecparsers: av1: uint8 range is not enough for av1_bitstreamfn_ns
Part-of: <https://gitlab.freedesktop.org/gstreamer/gst-plugins-bad/-/merge_requests/1464>

+++ gst-libs/gst/codecparsers/gstav1parser.c
@@ -293,8 +293,8 @@ av1_bitstreamfn_su (GstBitReader * br, guint8 n, GstAV1ParserResult * retval)
 /* 4.10.7
  *
  * Unsigned encoded integer with maximum number of values n */
-static guint8
-av1_bitstreamfn_ns (GstBitReader * br, guint8 n, GstAV1ParserResult * retval)
+static guint32
+av1_bitstreamfn_ns (GstBitReader * br, guint32 n, GstAV1ParserResult * retval)
 {
   gint w, m, v;
   gint extra_bit;
